add delete and openjdk Dockerfile

+++ src/main/java/com/auycro/score/repository/ScoreRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.auycro.score.entity.ScoreEntity;
 
-public interface ScoreRepository extends CrudRepository<ScoreEntity, Integer> {
+public interface ScoreRepository extends CrudRepository<ScoreEntity, Long> {
   
 }
